+++ Packages/com.yontalane.dialog/Runtime/AssemblyInfo.cs
@@ -1,0 +1,4 @@
+using System.Runtime.CompilerServices;
+
+[assembly: InternalsVisibleTo("Yontalane.DialogUGUI.Runtime")]
+[assembly: InternalsVisibleTo("Yontalane.DialogUIElements.Runtime")]